context menu container refactored
