+++ src/ui/context-menu-container/context-menu-container.types.ts
@@ -1,7 +1,6 @@
 export type ContextMenuProps = {
   children: React.ReactElement;
   menu: React.ReactNode;
-  contextMenuRef: React.MutableRefObject<ContextMenuRef | null>;
 }
 
 export type ContextMenuRef = {
