+++ src/ui/tab-button/index.ts
@@ -1,1 +1,2 @@
 export * from './tab-button'
+export * from './tab-button.types'
