+++ src/ui/tab-dropdown/index.ts
@@ -1,1 +1,2 @@
-export { TabDropdown } from './tab-dropdown'
+export * from './tab-dropdown'
+export * from './tab-dropdown.types'
