+++ src/ui/tabs-panel/index.ts
@@ -1,1 +1,2 @@
 export * from './tabs-panel'
+export * from './tabs-panel.types'
